+++ catroid/src/androidTest/java/org/catrobat/catroid/test/catblocks/ScriptSplitSingleTest.kt
@@ -55,7 +55,7 @@ class ScriptSplitSingleTest {
     private val projectName = javaClass.simpleName
 
     companion object {
-        private const val TIMEOUT: Long = (5).toLong()
+        private const val TIMEOUT: Long = 5
         private const val TEST_SPRITE = "testSprite"
     }
