+++ catroid/src/androidTest/java/org/catrobat/catroid/test/catblocks/ScriptSplitUserDefinedBrickTest.kt
@@ -55,7 +55,7 @@ class ScriptSplitUserDefinedBrickTest {
     private val projectName = javaClass.simpleName
 
     companion object {
-        private const val TIMEOUT: Long = (5).toLong()
+        private const val TIMEOUT: Long = 5
     }
 
     @get:Rule
