+++ catroid/src/androidTest/java/org/catrobat/catroid/test/catblocks/UserDefinedBrickTest.kt
@@ -54,7 +54,7 @@ class UserDefinedBrickTest {
     private val projectName = javaClass.simpleName
 
     companion object {
-        private const val TIMEOUT: Long = (5).toLong()
+        private const val TIMEOUT: Long = 5
     }
 
     @get:Rule
